Add initial support for loading of movie "star ratings"
The Sage API docs aren't clear on how to actually load this data so I'm
making a best guess effort here.  I'd be surprised if this actually
worked, but it's something to start with.  I've emailed Sage support
asking for the actual details on how to do this, time will tell if they
actually reply.

Fixes issue #344