Issue 351: Add support for setting of a show's language
The EPG plugin itself will now recognize and set a show's language if
its value is anything other than 'English' or null, which is assumed to
also be English.

The EPG plugin will also set a program's language based on the
descriptionLanguage value of a program.  So channels that, for example,
broadcast in French and also provide their descriptions in French will
now have their program's language value set to French.

Airing filters can also set its associated program's language now by
simply overriding the 'broadcastLanguage' property of the object within
the filter.  When that value is overridden, the EPG plugin will
recognize that after the filter executes and will reset the language in
the associated program.  This should simplify anyone's existing filters
greatly.

Airing generators CANNOT set its associated program's language at this
time.  An attempt to do so will generate a warning in the logs.  If
someone actually hits this and needs this functionality then I'll tackle
it then, otherwise it's way too much hassle for something that I doubt
will be often used by anyone.  The req'd fix to make this work for
airing generators is described in the comments of the code that would
need to be fixed.

Fixes issue #351